requirements.txt
beautifulsoup4
pandas
python-dotenv
rispy
tiktoken
weaviate-client
